third_party/WebKit/LayoutTests/fast/html/imports/encoding.html

<!DOCTYPE html>
<head>
    <script src="../../../resources/js-test.js"></script>
    <link rel="import" href="resources/no-encoding.html">
    <link rel="import" href="resources/meta-encoding.html">
</head>
<body>
<script>
jsTestIsAsync = true;
window.onload = function() {
    description("Test that UTF-8 decoding is applied appropriately to imports");

    links = document.querySelectorAll('link[rel=import]');
    debug("Default encoding should be UTF-8...");
    shouldBeEqualToString("links[0].import.characterSet", "UTF-8");
    debug("And a meta tag should not override it...");
    shouldBeEqualToString("links[1].import.characterSet", "UTF-8");
    finishJSTest();
};
</script>
</body>
