third_party/WebKit/LayoutTests/fast/html/imports/encoding-expected.txt

Test that UTF-8 decoding is applied appropriately to imports

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default encoding should be UTF-8...
PASS links[0].import.characterSet is "UTF-8"
And a meta tag should not override it...
PASS links[1].import.characterSet is "UTF-8"
PASS successfullyParsed is true

TEST COMPLETE
